# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/test-drives": {
      "post": {
        "description": "Logged-in: send carId, scheduledAt, notes. Guest: send name, phone, email, model, preferredDate, timeSlot.",
        "operationId": "TestDrivesController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Test drive scheduled"
          }
        },
        "summary": "Schedule test drive (emits test-drive.created event)",
        "tags": [
          "Test Drives"
        ]
      },
      "get": {
        "operationId": "TestDrivesController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of test drives"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List test drives",
        "tags": [
          "Test Drives"
        ]
      }
    },
    "/api/test-drives/stats": {
      "get": {
        "operationId": "TestDrivesController_getStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Test drive stats"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get test drive statistics",
        "tags": [
          "Test Drives"
        ]
      }
    },
    "/api/test-drives/{id}": {
      "get": {
        "operationId": "TestDrivesController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Test drive details"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get test drive by ID",
        "tags": [
          "Test Drives"
        ]
      },
      "patch": {
        "operationId": "TestDrivesController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTestDriveStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status updated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update test drive status (emits test-drive.status-changed event)",
        "tags": [
          "Test Drives"
        ]
      },
      "delete": {
        "operationId": "TestDrivesController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Test drive cancelled"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancel test drive",
        "tags": [
          "Test Drives"
        ]
      }
    },
...
  },
  "info": {
    "title": "Baccouche Motors API Gateway",
    "description": "API Gateway for Baccouche Motors microservices.\n\n## Event-Driven Design (RabbitMQ)\nThis system follows a strict **Event-Driven Design** pattern:\n- **Domain events** are emitted on create/update/delete operations\n- **Consumers** react asynchronously (emails, stats, audit)\n\n| Event | Producer | Consumers |\n|-------|----------|-----------|\n| `auth.forgot-password-requested` | Auth | Notifications (email) |\n| `car.created` / `updated` / `deleted` | Cars | Admin (stats) |\n| `news.created` / `updated` / `deleted` | News | Admin (stats) |\n| `contact.created` / `updated` | Contacts | Notifications, Admin |\n| `test-drive.created` / `status-changed` | Test Drives | Notifications, Admin |\n\n**Flow**: Gateway → RabbitMQ RPC (sync) → Microservice → emits to `baccouche.events` exchange → Consumers (Notifications, Admin)",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateTestDriveStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "confirmed",
              "completed",
              "cancelled",
              "rejected"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
...
    }
  }
}